refactor: renamed the directory resume-builder-v2 as resume-builder.
