+++ src/components/resume-builder/components/edit-panel/ActivitiesEditBox.tsx
