+++ src/components/resume-builder/components/edit-panel/EditPanel.tsx
