+++ src/components/resume-builder/components/edit-panel/SkillsEditBox.tsx
