+++ src/components/resume-builder/components/resume-preview/Activities.tsx
