+++ src/components/resume-builder/config/work-placeholders.ts
